refector: 모듈화 및 프로그래스바 기능 수정
