+++ style/style.css
@@ -224,7 +224,7 @@ main .news-list-wrap .progress-tab .progress-ratio {
   height: 40px;
   flex-shrink: 0;
   z-index: 0;
-  transition: width 0.5s;
+  transition: width 0.1s;
 }
 main .news-list-wrap .progress-tab .text-category-name {
   display: flex;
